requirements.txt
attrs==19.3.0
click==7.0
nltk==3.4.5
OpenNMT-Py==1.1.1
sentencepiece==0.1.83
textdistance==4.1.5
